requirements.txt
pandas>=2.0.0
openpyxl>=3.1.2        # Lecture des fichiers Excel
PyYAML>=6.0            # Parsing des fichiers YAML
tabulate>=0.9.0        # (optionnel) Pour jolis tableaux CLI si tu veux
